src/services/rag/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
qdrant-client==1.6.0
openai==1.3.5
sentence-transformers==2.2.2
pypdf==3.17.1
python-docx==1.1.0
beautifulsoup4==4.12.2
markdown==3.5.1
lxml==4.9.3
psycopg2-binary==2.9.9
prometheus-client==0.19.0
python-multipart==0.0.6
